+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2022-11-06
+### Added
+- Added Isekai Succubus Heritage (Tiefling Heritage).
+- Added Isekai Angel Heritage (Aasimar Heritage).
+- Added Overpwered Ability `Super Buff`.
+### Changed
+- Buffed Edge Lord's "Supersonic Combat" to allow for use in CMB and to qualify for feats that require strength.
+
 ## [1.4.1] - 2022-11-05
 ### Fixed
 - Fixed Drake Companion Selection.